ComRV: Parse grouping file
Read grouping info from csv file and store in a hash table indexed by
a hash of the function name.

riscv_test_ovl_grouping_table shows how to create and free the table.
This function can be removed once riscv_create_ovl_grouping_table is
integrated in the correct place.

At the moment, creating the table will be attempted regardless of
whether --grouping-file is specified or not. This is just for testing
purposes and will not happen once riscv_test_ovl_grouping_table is
removed.