ComRV: Add partial support for multigroups
This makes it possible for a symbol to exist in multiple groups.
The contents of the section associated with the symbol will
then be copied into all of the output group sections that it
exists in.

The multigroup table, and references into the multigroup table
are not yet materialized. Instead a placeholder value is produced.