ComRV: If no group file provided, put each overlay function in a separate group
In check_relocs we marks any symbols which are referred to with
an overlay relocation. When sizing dynamic sections we then
check if a grouping file was provided, if no grouping file was
provided then each symbol referred to by an overlay relocation
is placed in its own overlay group.

If a grouping file was provided, but it doesn't cover all of the
symbols referred to by an overlay relocation, then this will
also now error.

A stub has been added in size_dynamic_sections where we will
need to call out to the overlay grouping tool in future.