ComRV: Improve the handling of --grouping-tool-args option
Options are now correctly passed along to the grouping tool
when --grouping-tool-args is provided to the linker. The arguments
are comma separated, so the following:

  riscv-unknown-elf-ld --grouping-tool-args --arg1,--arg2,--arg3 ...

Results in the following command being executed to run the
grouping tool:

  comrv-group --arg1 --arg2 --arg3 -o tmp-symbol-grouping.csv
              tmp-symbols-to-group.csv

The -o option specifies the output file. After the output file name
input files containing symbols and sizes follows.

The command name "comrv-group", and the input and output filenames
"tmp-symbols-to-group.csv" and "tmp-symbol-grouping.csv" are
currently hardcoded.

The command "comrv-group" is executed by forking the "ld" process.
As a consequence it is *not* run from within a shell and must be
an executable.