ComRV: Avoid assigning/resolving symbols more than once
There were a couple of bugs which meant that auto-assigning of
symbols to overlay groups, and resolving the offset of a symbol
in groups could happen multiple times for the same symbol. This
happened because the symbol can appear multiple times in the
symbol tables of the input bfds, for example once for the
definition of the symbol, and another time for a (previously
undefined) reference to the same symbol.

The effects of these bugs were two fold:

* When auto-assigning, the same symbol would be assigned multiple
  times, resulting in multiple copies being copied to .ovlgrpdata
  at different offsets.
* When resolving the offset for a symbol in a group, the
  offset would be recalculated for each occurrence of the
  symbol (even if it was only assigned once, to one group). The
  size of the symbol would be added to the size of the group
  multiple times, and the next occurrence of the symbol would
  cause its offset to be recalculate dto the end of the now
  resized group.

The solution to these problems is to have a flag to note both when
the symbol has been assigned, and when the symbol has been resolved,
so these operations only happen once. In addition any occurence of
a symbol which is not a definition is ignored.