ComRV: Error if an overlay symbol is referenced to by a call reloc
This checks whether a symbol is referred to by both a normal
call relocation as well as a overlay call relocation, and
if so reports this to the user. It also checks whether a symbol
which is assigned to an overlay group is referred to by a
non-overlay relocation, and reports this as an error.