ComRV: Add an --overlay-tool option, also tweaks to --overlay-tool-args
The --overlay-tool option is used to provide the command to invoke
the grouping tool from the linker.

The linker is now able to understand two options which are provided
as an argument to the --overlay-tool-args option. These are the
--in-file and --out-file options, and the arguments to these options
specify the input .csv file containing symbols to be grouped and
their sizes, and the file that the grouping tool will write symbol
groupings to respectively.

These --in-file and --out-file options are parsed by the
linker, but they are also passed through the grouping tool itself.