gdb: Initial skip overlay manager implementation
This is to allow the overlay manager to be skipped when stepping into or
out of a function. While stepping, if we detect that the overlay manager
has been entered, an internal breakpoint is set at the exit point of the
overlay manager. We then continue until this breakpoint and step over it
to reach the start of the called or returned to function.