gdb: Support line information lookup and disassemble command
In this commit initial support is being added for looking up line
table information and function start/end addresses based on the
current $pc, even when the $pc value is within a mapped overlay.

There is also support for the disassemble command, when a function is
currently mapped in.

Initial support for single stepping (by line) through overlay code is
in place, though stepping over calls into the overlay manager (with
'next') will not work as backtracing doesn't currently work.

As with the previous commit, this code is still very rough, and should
be considered work in progress.