ComRV: Replace assertion in ovloff with error
Previously an assertion would be triggered when trying to
materialize an overlay token for a symbol which is not
allocated to any overlay groups. This can happen for undefined
symbols, since it is possible for them to be referenced without
them being assigned to an overlay group (because the autogrouping
only deals with symbols which are defined).