gdb: Remove hard-coded addresses related to overlay management
Use the Python overlay manager to figure out the regions in which
various parts of the overlay system are stored (the storage area and
the cache area) then pass this information into GDB core, and use this
information for managing breakpoints.  This replaces some hard coded
address masks that I was previously using, and should make the current
code a little more resilient int the face of different test programs.