ComRV: Allow custom CRCs
This forks the CRC implementation already used by the linker, and adds
command line arguments for all the CRC32 paramters, such that other
CRCs can be enabled at link time without recompiling the the linker.

Default settings for the CRC are stored in the riscvelf emultempl, these
can be changed once it is agreed what the values should be going
forward.

Currently this has a dependency on a fork of libibery's crc32.c named
elfnn-riscv-crc.inc. It should be discussed where this code should live
long term, since it is it not really RISC-V specific.