gdb: Delay loading region data until we really need it
If the Python overlay manager is loaded before the ELF to debug then
we don't have access to the region information.  Currently we were
trying to setup this information from the constructor of the Python
overlay manager object.  This means that in the case where the Python
is loaded before the ELF no regions (source or cache) were setup, and
GDB was unable to correctly map breakpoint addresses.

This commit delays loading the region data until we first need it,
which should be when we try to insert a breakpoint, which should be
after the ELF is loaded, at which point the region information should
be available.