gdb: Don't disable use of Z0/z0 packets
Initially I disabled the use of Z0/z0 packets.  The reason for this is
that on the HiFive1 board I was being prevented (by OpenOCD maybe?)
from placing more than 2 breakpoints into the overlay region using the
Z0/z0 packets.  However, if I disabled use of Z0/z0 and allowed GDB to
fallback to using the older software breakpoint technique (direct
write to memory, managed from GDB) then I could place as many software
breakpoints into overlay code as I liked.

However, on the WD FPGA board disabling use of Z0/z0 causes other
problems.  If we are not using the Z0/z0 packet, then it seems as
though the software breakpoint instruction almost gets "stuck" in the
processor pipeline cache (or similar).  Despite GDB trying to remove
the breakpoint we continue to hit a breakpoint at the current $pc.

Switching to use of Z0/z0 means we can make more progress on the WD
FPGA target.  On the HiFive1 board we should disable use of Z0/z0
outside of this script.