gdb: handle multiple breakpoints for the same overlay location
It was discovered that if the user places two breakpoints at the same
address, and the address is an overlay location, then GDB would get
into a weird state from which it could no longer correctly insert the
overlay breakpoints.

The problem was that the insert/remove overlay breakpoint location
code, only updates the address field on the first breakpoint, all
duplicates (i.e. at the same address) were left alone.

So, if we have:

  BP 1 @ storage-address-1, inserted = no
  BP 2 @ storage-address-1, inserted = no

Now, the mapping is set up so that:

  storage-address-1 =maps to=> cache-address-1

When we next insert the breakpoints we get this:

  BP 1 @ cache-address-1, inserted = yes
  BP 2 @ storage-address-1, inserted = no

As GDB saw both breakpoints with storage-address-1, and only inserted
the first, however, during the insert we updated the address to the
cache address.

Assuming the mapping remains unchanged, GDB stops, and removes all
breakpoints, we now have this:

  BP 1 @ cache-address-1, inserted = no
  BP 2 @ storage-address-1, inserted = no

GDB leaves the address field changes so the user can see where the
breakpoint would/is going to be inserted.  Now when we continue GDB
fails to spot that the breakpoints are really for the same address,
and so we end up like this:

  BP 1 @ cache-address-1, inserted = yes
  BP 2 @ cache-address-1, inserted = yes

This isn't good.  Now, assuming the mapping still hasn't changed, we
stop, and move to this state:

  BP 1 @ cache-address-1, inserted = no
  BP 2 @ cache-address-1, inserted = no

Now we continue again, this time the addresses of the two breakpoints
are the same, so GDB understands they are for the same location, and
so we get this:

  BP 1 @ cache-address-1, inserted = yes
  BP 2 @ cache-address-1, inserted = no

Now, we stop.  However, this time the mapping has changed to this:

  storage-address-1 =maps to=> **nothing**

After stopping we see this:

  BP 1 @ storage-address-1, inserted = no
  BP 2 @ cache-address-1, inserted = no

GDB reverted the addresses on the breakpoint it removed, but BP 2 was
not removed this time, as a duplicate it was never inserted.  So the
breakpoint remains in a state that appears to be mapped in.

Some time later, the mapping changes to this:

  storage-address-1 =maps to=> cache-address-2

Now we continue, again GDB fails to see that the two breakpoints are
really for the same address, and so tries to insert them both,
however, BP 2 already appears to be mapped in to a different location,
this should have been fixed at removal time, so GDB throws an error.

The solution is to have GDB update all breakpoint locations at the
same address when inserted and removing overlay breakpoints.  With
this, after the first insert phase we now see:

  BP 1 @ cache-address-1, inserted = yes
  BP 2 @ cache-address-1, inserted = no

Both locations have the updated address, but still only one is
inserted.  When the mapping changes and we remove BP 1, we now also
update both locations, so afterwards we go back to:

  BP 1 @ storage-address-1, inserted = no
  BP 2 @ storage-address-1, inserted = no

Everything is now consistent.