gdb: First phase of ComRV/Overlay Manager cleanup
This is mostly about renaming the API into the overlay manager so that
the naming scheme is a little more consistent, and inline with how
things are named in the other tools.

However, switching over to the new API was a good opportunity to go
through the breakpoint code and give it a bit of a polish, improving
error messages and debug output, and removing some redundant code.

The only functional change after this commit should be changes to
debug output and error message content, nothing else should change.