gdb: Allow unmapped s/w breakpoints to be removed
In the case where we have a s/w breakpoint placed into an overlay,
which is evicted, and replaced with a new overlay, GDB needs to be
able to safely remove the s/w breakpoint.

The reason is that OpenOCD will give an error if we try to insert a
breakpoint at an address where a previous s/w breakpoint has not been
removed (even though it shouldn't).

The solution in this patch is that GDB will read the memory where the
s/w breakpoint used to be (this will now have been modified by the new
overlay), then tell the target to remove the s/w breakpoint, this will
likely corrupt the memory in the new overlay.  Finally, GDB writes
back the original memory contents (those from before the s/w
breakpoint was removed), this should leave things as they are supposed
to be.

Initial testing on FPGA looks good.