gdb: Resort the locations list after updating addresses
When we remove a breakpoint location we restore its address to the
storage area address.  We do this inside a loop controlled with the
macro ALL_BP_LOCATIONS_AT_ADDR.  The problem is that this loop relies
on the locations being sorted by address in order to quickly get the
loop started.

If we are removing multiple locations (say many breakpoints are inside
an overlay that just got unmapped) then the first breakpoint will
remove fine, and its locations will be updated.  But, once it is
removed the location list is no longer sorted, as a result, when we
try to remove a second breakpoint, we _might_ find that the
ALL_BP_LOCATIONS_AT_ADDR loop doesn't start correctly, and we don't
update the corresponding breakpoint locations.  This will then leave
some breakpoints in a state where they appear to be inserted, even
though their corresponding overlay is unmapped.

Fix this issue by re-sorting the location list any time we have to
adjust some breakpoint locations.