gdb: Fix lookup of comrv_ret_from_callee symbol
Everest's Eclipse setup uses -add-symbol-file to load the executable
instead of the more usual -file-exec-and-symbols.  This then exposed
an issue in the ComRV unwind code, where we were only searching the
main executable for the symbol comrv_ret_from_callee.

Fixed this by searching all objfiles for the symbol, rather than just
the main executable.