gdb: Move some of the stack unwind logic into Python code
In an ideal world all of the stack unwind changes would be implemented
as a Python extension.  Unfortunately, GDB doesn't have the
infrastructure to support this (yet).

In this commit I move the core of the unwind algorithm out of GDB and
into the Python code.  The benefit of this is that this code can now
be tweaked and improved without having to recompile GDB.  If the
algorithm for unwinding changes that it should be possible for this to
be updated without having to ship a new version of GDB.