gdb: Avoid reference counting error when reading from lists
A bug in our use of the Python API meant that some objects were having
their reference count incorrectly decremented twice.  When accessing
objects within a list with PyList_GetItem, we don't get a new
reference, instead we borrow the reference from within the list, so we
shouldn't decrement the reference when we're done with it.