gdb: Don't assume ordering of multi-group members
Previously it was assumed that the first overlay group in a
multi-group would be the one with the debug information.  However,
there was no guarantee that this would always be true.

This commit removes this assumption.  GDB now searches for the overlay
group member that has debug information available.  It is still
assumed that there will only be one member (and exactly one member)
that has debug information available, anything else will result in an
error being thrown.