gdb: Initialise Python overlay manager once ComRV B/P is hit
Delay initialisation of the Python overlay manager until the ComRV
event breakpoint has been hit at least once.  This should avoid issues
where the machine state appears to indicate that ComRV is initialised,
when really it isn't.

After this commit a global flag is set to zero when the script is
loaded, and incremented when the event breakpoint is hit.  There's
currently no way to "reset" the script back into a state indicating
that ComRV is not initialised, but I think that due to other state
being cached, even resetting the counter would not be enough, so I've
not worried about that for now.