gdb: Remove use of g_stComrvCB.ucTablesLoaded symbol
Now that we are check for the first hit of the overlay event
breakpoint in order to decide if ComRV is initialised or not, then we
no longer need the g_stComrvCB.ucTablesLoaded symbol.

In order to allow Everest to remove this when convenient, remove use
of this symbol from our code.