gdb: Convert overlay breakpoints to their storage address
When a breakpoint is placed into the cache region using a cache
address directly, then convert this into a storage area address.

This comes up when using 'next' to skip over a call to an overlay
function, from within an overlay function.

So, initially we are in an overlay function, and the user invokes the
next command.  GDB steps forward until it enters ComRV, at which point
we find the returns address in the previous frame (the original
overlay function) and place a breakpoint there.

The problem is that the return address from the ComRV entry point to
the previous overlay function will be a cache address, so the next
resume breakpoint will be placed at a cache address.

GDB identifies this as an overlay breakpoint, but understands that
this breakpoint has no valid storage address, and throws an error.

The solution is that the cache only address is converted to a storage
area address.  This should be fine, at the point the breakpoint is
created the overlay containing the original function will still be
mapped, so all should be good.

If the user tries to place a breakpoint within the cache region, and
there's no mapped overlay for that cache address then GDB will throw
an error and refuse to give the breakpoint a storage area address.