gdb: Fix function start address when comrv skipping
When stepping by line into an overlay function with comrv skipping on,
fill_in_stop_func was incorrectly calculating the function start address
as the start of the overlay cache.

Now, if a step-resume breakpoint is hit and the stop_pc is an overlay
cache address, change stop_pc to the storage address so that the debug
info is used to fill in the correct function start.