ComRV: Fix offsets in mapfile
The mapfile was the last piece of code relying on the (unrelaxed)
offsets which should only be used for sorting. This has been replaced
with a new processed_offset that is calculated relocation resolution.

An assert has also been added to check that this value has been filled
in; if it triggers then a more costly offset lookup would be needed.