Rewrite the ComRV stack unwinding
This commit rewrites the ComRV stack unwinding.  We no longer hide the
ComRV stack frames from the user, so an example stack will now look
like this:

  #0  0x80000c24 in OverlayFunc3 () at .../WD-Firmware/demo/demo_comrv_baremetal_multigroup.c:71
  #1  0x20401072 in comrv (token=1179655) at .../WD-Firmware/comrv/comrv_entry.S:164
  #2  0xe0000456 in OverlayFunc2 () at .../WD-Firmware/demo/demo_comrv_baremetal_multigroup.c:78
  #3  0x20401072 in comrv (token=541067496) at .../WD-Firmware/comrv/comrv_entry.S:164
  #4  0xe0000258 in OverlayFunc1 () at .../WD-Firmware/demo/demo_comrv_baremetal_multigroup.c:85
  #5  0x20401072 in comrv (token=2147483649) at .../WD-Firmware/comrv/comrv_entry.S:164
  #6  0xe0000220 in OverlayFunc0 () at .../WD-Firmware/demo/demo_comrv_baremetal_multigroup.c:93
  #7  0x20401072 in comrv (token=3) at .../WD-Firmware/comrv/comrv_entry.S:164
  #8  0x20400396 in demoStart () at .../WD-Firmware/demo/demo_comrv_baremetal_multigroup.c:111
  #9  0x20400322 in main () at .../WD-Firmware/demo/main.c:72

There is a new option a user could use:

  (gdb) set comrv show-frames off

Which will cause the 'comrv ()' frames to be hidden in the backtrace,
unless the 'comrv' frame is the currently selected frame (as a
backtrace that doesn't show the currently selected frame is very
confusing).

The 'token' parameters can also be disabled with:

  (gdb) set comrv show-tokens off

The token parameters show the value that was passed in $t5 for each
comrv frame and is either pulled from $t5 or fetched from the ComRV
stack.

The nice thing about this rewrite is that it is now entirely written
in Python with no changes required to core GDB.  As a result I'm able
to remove some core GDB changes, and delete some unneeded Python
support code.

The underlying algorithm for unwinding is completely unchanged
though.