Rewrite ComRV skipping to work with RTOS
The comrv engine previously had two entry points, each with a unique
exit point:

  ENTRY POINT              EXIT POINT
  comrvEntry               comrv_invoke_callee
  comrv_ret_from_callee    comrv_exit_ret_to_caller

In the previous skipping approach, if either entry point was hit during
a step, then a bp would be enabled on the corresponding exit point. As
the engine is in read-only memory, this would be a hw bp. To satisify
the requirement that comrv uses a maximum of 1 hw bp at any time, the
overlay event bp was disabled whilst skipping through then engine,
however, for RTOS programs, the overlay event bp must be enabled at all
times.

In this new approach, if an entry point is hit during stepping, a
step-resume bp is set in the destination function. If the entry point is
comrvEntry, the bp is set at the start of the callee. If the entry point
is comrv_ret_from_callee, the bp is set on the return address in the
caller. If the destination is in RAM, this will be a sw bp, but if it is
in ROM, it will be a hw bp (thus exceeding the 1 hw bp). This is ok, but
needs to be documented for the user.

The following two new entry points are now also supported, if the user
steps into these from a context switch, the comrv engine will be skipped:

  ENTRY POINT                             EXIT POINT
  comrvEntry_context_switch               comrv_invoke_callee
  comrv_ret_from_callee_context_switch    comrv_exit_ret_to_caller

If the user steps into the comrv engine anywhere other than the four
defined entry points, skipping will not occur.

Skipping now happens for `step` but not `stepi`.

The skip handling now takes place predominantly in overlay.c and the
python overlay manager rather than breakpoint.c. There are still a few
small changes to infrun.c.