Update calculation of multi-group index size on the ComRV stack
GDB was making invalid assumptions about how to calculate the
size of the multi-group index field on the comrv stack. This commit
uses the new g_uiComrvInfo symbol from comrv to calculate the size
correctly.

The lower nibble of g_uiComrvInfo contains the multi-group index offset.
If the offset is 11, the multi-group index size is 1 byte. If the offset
is 14, then the multi-group index size is 2 bytes.

comrv_stack_frame is now initialised with mg_index_offset instead of
is_mg, which can be inferred as mg_index_offset > 0.