gdb/comrv: Avoid calling value_of_register from pending frame
Work around for bug #25.

This is a temporary work around for bug #25.  The problem is that
value_of_register relies (in some cases) on having the frame-id for
the current frame, specifically, when the frame we're in is inlined,
then we need to have the frame-id for the current frame.

GDB only switch to use value_of_register here to handle user
registers (a.k.a. pseudo registers), which luckily we don't make use
of on RISC-V (yet) so this switch back will resolve the issue without
introducing any regressions.

However, if we want to upstream this fix then we'll need a better
solution long term.