gdb/riscv: Dump/load CSRs based on target description
The CSRs will be dumped in the order they appear in the target
description.

The target description is now dumped to the corefile. When loading the
core file, the target description is read from the corefile and used to
determine how the CSRs are stored in the corefile in order to load them.