gdb: Don't add last entry to the disassembly range if it is an END marker
In the previous loop we already skip most end of sequence markers,
however, if we reach the end of the line table, and still haven't
quite reached our required high then currently we just lookup the SAL
for the $pc of the previous END marker, and use its .pc as our end
value.

The problem we're seeing is that the address coming back is in the
overlay storage region.

What appears to happen is that we have something like this going on.

| Address | foo.c | bar.c |
|---------+-------+-------|
|   0x100 | X     |       |
|   0x102 | END   |       |
|   0x104 |       |       |
|   0x106 |       |       |
|   0x108 |       | X     |
|   0x10a |       | x     |
|   0x10c |       | END   |
|   0x110 |       |       |
|   0x112 |       |       |
|   0x114 | X     |       |
|   0x116 | X     |       |
|   0x118 | END   |       |
|---------+-------+-------|

We request disassembly from 0x10a -> 0x110.  Once we leave bar.c's
line table, GDB tries to figure out a SAL for 0x110, at which point it
comes back with a SAL 0x102 -> 0x114, and so it uses 0x114 as the end
address for disassembling.

I considered capping the end address (from find_pc_line) at the
disassembly high address in do_mixed_source_and_assembly_deprecated,
but given the earlier loop already skips END markers, I don't think
skipping END markers here is a big issue.

Further I suspect that this end block is now completely redundant, as
I think you can _only_ get here for the last line in the line table,
which is guaranteed to always be an END marker.