gdb/comrv: Fix invalid cast error
If no executable has been set, GDB sets the default void pointer length
to 8 bytes. If a 32-bit executable is then loaded, the void pointer
length changes to 4 bytes, but comrv_unwinder.void_ptr_t was not being
updated to reflect this, which caused an invalid cast error.

To keep comrv_unwinder.void_ptr_t in sync with GDB, it is now updated in
comrv_unwinder.__call__.