ComRV: Fix support for GC sections
An assertion is triggered if a function has been placed into the ComRV
management tables that is then gc sectioned should the removal of that
function cause the containing group to change size. Rather than
re-calculate the sizes of these, make the linker aware of whether a
overlay function has been gc-section'd in the first place and not set
up any of the group data for this function.

As a side effect this can result in the number of groups being reduced,
since if only gc-section'd functions were explicitly placed into a
group, it is now regarded as empty, and another auto-placed function
can be placed into this location.