gdb/comrv: Hide some warnings about breakpoint address adjustment
When GDB hits a breakpoint, if the address of the breakpoint has been
adjusted from the original requested address then a warning is
printed.  Normally, this is a good thing, but when we consider overlay
breakpoints this is not so helpful.

Overlay breakpoints are (almost) always requested in the storage
region, and then adjusted into the cache region.  In this case, if we
continually spam the user with warnings then this can get annoying.

This commit disables the adjustment warning for overlay breakpoints,
if the current breakpoint location address is the matching cache
address for the requested storage region address.