gdb/comrv: new 'comrv groups LOC' command
Add a new command 'comrv groups LOC' that takes a location specifier,
as you would pass to the 'break' command, and reports which overlay
groups that location is in.

The output can look like this:

  (gdb) comrv groups main
              Group  Group       Group       Group
  Address     Number Start       End         Size    Offset
  0x2040031c      ** is not in overlay storage area **
  (gdb)

This indicates the location is not in any overlay groups.  Then, if
the location is in a multi-group we might see this:

  (gdb) comrv groups OverlayFunc1
              Group  Group       Group       Group
  Address     Number Start       End         Size    Offset
  0xe0000234  1      0xe0000200  0xe0000400  0x200   0x34
  0xe0000400  2      0xe0000400  0xe0000600  0x200   0x0
  (gdb)

This indicates that OverlayFunc1 is in two overlay groups.  A location
that is not in a multi-group will only produce a single line of output.