gdb/comrv: take care where we sort the breakpoint locations list
We should not sort the breakpoint locations list while inside an
iterator over the same locations list.  The iterator works by assuming
a list of pointers arranged sequentially in memory, if we sort this
list while iterating then an unprocessed element can be shuffled
earlier in the list and avoid being processed altogether.

Now we defer sorting until the iterator has completed.  This does mean
that there is now a short window of time in which the breakpoint
locations list might be unsorted.

The consequence of having an unsorted breakpoint location list is that
we must not use ALL_BP_LOCATIONS_AT_ADDR, this macro assumes that the
locations list is sorted in order to find the starting location in the
list, so instead we must just use ALL_BP_LOCATIONS, and manually skip
over non matching elements.