gdb/comrv: iterate over multi-group members, not multi-group tokens
In the Python overlay manager, make use of the new multi-group members
information to iterate over members, rather than tokens, in all
cases.  The advantage is that this gives us direct access to the
underlying overlay group object, without having to look it up.

In order to facilitate this transition I've added the id number to the
overlay group object, so each overlay group now knows what its id
number is (that is the index into the overlay groups table).