gdb/comrv: Don't error if the python unwinder return optimised out regs
Upstream GDB does not like it if the Python unwinder returns optimised
out registers (or unavailable registers).  This means that every
unwinder should be checking that the value is fully available before
calling add_saved_registers.

There are two problems here:

  1. There is no Python API to check values are 'available' as opposed
  to 'optimised-out'.  I guess the only option is to try and print the
  value in GDB and see what GDB prints.  If it's '<unavailable>' then
  you know.... this sucks.

  2. Having to wrap every call to add_saved_registers sucks, it's just
  asking for some calls to be missed which leaves us with a latent
  bug.

The right solution here is to allow any value to be thrown at
add_saved_registers, and just have GDB handle the values as needed.

This patch provides an minimally invasive solution to this problem.
It's possible that there's a better approach we should implement for
upstream.