gdb/comrv: prevent reads from comrv storage area
When a breakpoint is created GDB will read the location of the
breakpoint in order to figure out if there is already a breakpoint
instruction at this address - a "program breakpoint" or "permanent
breakpoint".

This check is carried out before the current overlay logic that
translates storage addresses to cache addresses, and as a result we
try to read from the storage area.

My solution for this is to just to short cut this function and return
false (indicating no program breakpoint) if the address being read is
in the storage area.

The only slight issue here is that if there _is_ a program breakpoint
at this address GDB will not spot this, and weird things might happen
in future.

However, there really is no way to avoid this.  In general GDB can't
see the storage area, so we have no way of knowing the contents of the
storage area until the address has been mapped into the cache region.

On the plus side, program breakpoints are a very specialist, and
seldom used feature, so it doesn't feel like too much of an issue to
just say that these can't be used within a ComRV overlay function.