gdb/comrv: allow Python overlay manager to adjust event bp location
Currently creating the overlay breakpoint location happens as soon as
GDB loads the executable, which is often before the Python overlay
manager is loaded.  As a result GDB will use the default fallback
overlay breakpoint location.

When the Python overlay manager is loaded, GDB does not currently
reset the event breakpoint location, so if the Python code uses a
non-standard event breakpoint location then GDB will ignore this if
the Python code is loaded after the executable.

This commit fixes this so that when the Python overlay manager is
loaded GDB will delete the overlay event breakpoint, and then recreate
the breakpoint, using the new location name.