gdb/comrv: prevent infinite loop in `comrv stack` command
During the comrvEntry code, while the comrv stack is being adjusted,
there is a point in time where the new stack frame has been allocated,
but the offset to the previous stack frame is still 0.  If at this
point the user runs the `comrv stack` command GDB will enter an
infinite loop as the zero sized stack frame means no progress is made.

This commit just causes GDB to break out of its loop early in this
case.