gdb/comrv: better lookup of the comrv token during comrvEntry
This commit improves the logic used to lookup the comrv token when
creating pseudo arguments for the comrv function frames.

In this commit I adjust the label used as the cut off for when we
expect to find the token in register $t5, and when we expect to find
the token on the comrv stack.