gdb/comrv: refactor helper functions for pulling tokens apart
Tweak the helper functions that are used to pull tokens apart.  We
previously only had functions whose names implied they should operate
on multi-group tokens, when in fact the functions were used for both
multi-group and non-multi-group tokens.

After this commit we have functions that operate on multi-group
tokens, and functions that operate on non-multi-group tokens.  The
functions each assert that they have the correct token types.  I've
then updated the rest of the python code to make use of the right
functions.

Finally, while updating which helper function was called I extended
some of the comments.