gdb/comrv: Reduce work done at overlay event breakpoint
When we hit the overlay event breakpoint we currently remove all
breakpoints and then reinsert all breakpoints.  This remove/insert
cycle allows GDB to handle overlay breakpoints that are in overlays
that are no longer mapped, or that have just become mapped.

However, this code is not as efficient as it could be.

Ideally we would only remove breakpoints when an overlay group was
unmapped, and only call insert for breakpoints that are in overlay
groups that have just become mapped.  Additionally overlays that have
"moved" (not something that I think ComRV does right now) would need
to be spotted and handled.

However, this ideal requires more analysis of the mapping, so
initially, to keep GDB simple, we simple removed all breakpoints, and
then inserted all breakpoints (unmapped overlay breakpoints are
silently ignored by the insert logic).

This simple approach is simple, but is inefficient, as non-overlay
breakpoints are removed and inserted each time we hit the overlay
event.

This commit improves things a little.  We still aren't at the "ideal"
as this would require much smarter analysis of the overlays, but we
can very easily figure out if a breakpoint is an overlay breakpoint or
not.  So that's what I do here.

When we hit the overlay event breakpoint remove only overlay
breakpoints, then reinsert all breakpoints.  As before, trying to
insert an unmapped overlay breakpoint will be ignored, similarly,
trying to insert an already inserted breakpoint is ignored.  This
change means only mapped overlay breakpoints will be removed and
inserted when we hit the event breakpoint, reducing slightly the work
that gets done at each event.