gdb/comrv: Don't use the skip comrv code for next like commands
When the user issues a next or nexti command we should not engage the
skip comrv code, we should instead just make use of the standard
next (or nexti) handling so that we return to the caller.

This commit just disables the skip comrv handling for next and nexti
commands.