gdb/comrv: fix skip-ovlymgr when using next to return from a function
After this commit:

  commit dd698a9d15c62edb4fcea62ee091ae0cc07eda48
  Date:   Mon Aug 17 17:43:21 2020 +0000

    gdb/comrv: Don't use the skip comrv code for next like commands

Using "next" to leave a function when 'set skip-ovlmgr 1' is also set
would no longer work, GDB would stop inside ComRV.

This is because after the above commit the skipping logic is no longer
applied for "next" or "nexti".  This was an overly aggressive change.

When moving from the caller to the callee with "next" it was correct
to not apply the skip comrv logic, and instead rely on GDB's normal
next handling to return us back to the caller, but, when using next to
return from the callee to the caller we should still be applying the
skip comrv logic.

This commit adds an extra condition to the skip comrv handling code.
If we are performing a next like command and the caller frame is not
where we started the next then we should apply the skip comrv logic,
otherwise (when we are performing a next, and the caller frame is
where we started the next) then we should not perform the skip comrv
logic (as GDB will naturally place us back in the caller).