gdb/comrv: use single byte load/store when removing s/w b/p
When removing a software breakpoint that is within an overlay that has
been unmapped we currently use a 4-byte load/store to preserve the
target memory.

An issue was reported where on some targets memory corruption appeared
to be creeping in, and the last place the memory appeared to be
touched was in this code.

I wondered if the target didn't like the miss-aligned 4-byte
load/store, and so in this commit I switch to using 4 1-bytes
loads/stores to see if this resolves the issue.

This code is fine to leave in long term, but also, if the issue is not
resolved by this change maybe I should revert this change.