gdb/comrv: allow reading overlay storage directly from the ELF
This commit is an initial attempt at adding support to GDB for reading
the overlay storage area directly from the ELF.  Reads to the storage
area are detected in the target memory reading code and redirected to
read from the ELF instead of allowing the read to go through to the
actual target.

In order to support this I had to allow GDB to load information about
all sections, including non-allocatable sections, into its section
table.  I have tried to go through GDB and in those places where I
think that we rely on only the allocatable sections being in the
table, I've added an extra check for the SEC_ALLOC flag, however, it
is possible that I've missed something.

My biggest worry is that should a user try to read otherwise invalid
memory, say at address 0, we might now start reading non-allocatable
debug sections that exist at address 0.

However, the benefit from this commit is that we can now disassemble
non-mapped overlay code, and (hopefully) the remote target should no
longer be seeing reads of invalid addresses related to overlay code.